config: Add a way to ignore controls
The Android HAL seems quite resilient to controls in the config that
don't actually exist or can't be toggled. UCM fails if setting any
control fails, so a mechanism to ignore some controls allows us to work
around that without having to change the XML or manually edit the
generated config.